Segmentation/requirements_2.txt
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
idna==2.10
joblib==1.0.1
jsonpatch==1.32
jsonpointer==2.1
kiwisolver==1.3.1
matplotlib==3.4.1
numpy==1.20.2
opencv-python==4.5.1.48
Pillow==8.2.0
pyparsing==2.4.7
python-dateutil==2.8.1
pyzmq==22.0.3
requests==2.25.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.16.0
threadpoolctl==2.1.0
torch==1.8.1
torchfile==0.1.0
torchvision==0.9.1
tornado==6.1
tqdm==4.60.0
typing-extensions==3.10.0.0
urllib3==1.26.4
visdom==0.1.8.9
websocket-client==0.59.0